organize pod msg parsers into a folder
