+++ parsers/pod_msg/parse_06.py
